Create Essie animation
In this example, we have a simple red square (represented by the `box` class) that moves from left to right and changes its background color to blue in the middle of the animation. The animation is defined using CSS keyframes.

To use this animation, save the HTML code in an HTML file (e.g., `index.html`), the CSS code in a CSS file (e.g., `styles.css`), and the JavaScript code in a JavaScript file (e.g., `script.js`). Make sure all files are in the same directory, and open the HTML file in a web browser to see the animation in action.

Feel free to modify the CSS properties and keyframe percentages to create different animations. You can also add additional elements and apply different animations to them.